refactor(agents-desktop): extract runtime entry helpers
Move runtime entry projection and MCP snapshot helpers into runtime modules while leaving lifecycle orchestration in main.ts.
